Broke out title animation functions and fixed a bug with animate charge by passing in game object. #28
